requirements.txt
openai
python-dotenv
langchain-openai
langchain-community
beautifulsoup4
faiss-cpu
langchainhub
langsmith
langgraph